prospect/requirements.txt
# Core dependencies
numpy
faiss-cpu
sentence-transformers
pydantic>=2.0.0
loguru>=0.7.0
rich>=13.0.0
python-dotenv>=1.0.0
click>=8.0.0

# LLM clients
google-generativeai>=0.3.0
openai>=1.0.0
google-adk>=0.1.0

# External intelligence and APIs
tavily-python>=0.3.0
requests>=2.31.0

# Data processing
aiohttp>=3.8.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
playwright>=1.40.0
numpy>=1.21.0 # For RAG

# RAG and Embeddings
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0 # Use faiss-gpu if GPU is available and preferred

# MCP Server dependencies
flask>=2.3.0
sqlalchemy>=2.0.0
gunicorn>=21.0.0
celery>=5.4.0
redis>=5.0.4

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
mypy>=1.0.0

# Optional dependencies for future features
# langchain>=0.1.0
# crewai>=0.1.0
